Making this middleware compatible with botpress-analytics
Botpress analytics requires a populated users database as it relies on the IDs there as foreign keys.

This fucking sucks, but to enforce this, we're calling db.saveUser() on every incoming message to make sure the user is in the DB before botpress-analytics tries to save it's record.

db.saveUser() is not documented, but from what I read in the code it's not updating existing records.

THIS FUCKING SUCKS (but at least it works)